add tools
The xipfstool was in the kernel directory and built there, but it is only used
by the user-space applications and makes more sense being built here in
userland. The only reason it was in with the kernel code is because it
requires a header file from the kernel. That requirement is now met in the
Makefile with an environment define: FROSTED_KERNEL.

Set FROSTED_KERNEL to the base directory of the FROSTED kernel, then you can
build the userland applications (which need the xipfstool).

Signed-off-by: Trevor Woerner <twoerner@gmail.com>